refactor: mv shipbuildsystem
