+++ src/app/information/components/guides/shipBuilds.ts
@@ -1,1 +1,1 @@
-export { USCBuildSystem as default } from '../../ship-builds/components';
+export { USCBuildSystem as default } from '../../ship-builds/uscBuildSystem';
