--- src/app/information/ship-builds/components/index.ts
@@ -1,1 +1,0 @@
-export * from './uscBuildSystem';
